pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "miniqa"
version = "0.1.0"
description = "An easy to use automated QA testing tool based on QEMU"
readme = "README.md"
requires-python = ">=3.9"
license = "GPL-3.0"

dependencies = [
    "Hypercorn>=0.18.0",
    "numpy>=2.0.0",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "ruamel.yaml>=0.19.1",
    "Quart>=0.20.0",
    "tqdm>=4.67.3",
    "websockify>=0.13.0",
]

[project.optional-dependencies]
default = []  # to simplify Containerfile
ocr     = ["ppocr-lite[fast]>=0.5.5"]

[tool.setuptools.packages.find]
include = ["miniqa*"]

[tool.setuptools.package-data]
miniqa = ["**/*.js", "**/*.css", "**/*.html"]
